refactor(l2): move deployer and system_contracts_updater to ethrex binary
