+++ cmd/ethrex/.gitignore
@@ -1,0 +1,2 @@
+# Dotenv file
+.env
