+++ cmd/ethrex/Cargo.toml
@@ -21,9 +21,11 @@ ethrex-sdk.workspace = true
 ethrex-l2-rpc.workspace = true
 
 bytes.workspace = true
+genesis-tool = { path = "../../tooling/genesis" }
 hex.workspace = true
 tracing.workspace = true
 tracing-subscriber.workspace = true
+thiserror.workspace = true
 k256.workspace = true
 clap = { workspace = true, features = ["string"] }
 clap_complete.workspace = true
@@ -98,3 +100,5 @@ harness = false
 
 [build-dependencies]
 vergen-git2 = { version = "1.0.7", features = ["rustc"] }
+ethrex-sdk.workspace = true
+hex.workspace = true
