+++ cmd/ethrex/l2/deployer/mod.rs
@@ -1,3 +1,6 @@
+mod errors;
+pub mod options;
+
 use std::{
     fs::{File, OpenOptions, read_to_string},
     io::{BufWriter, Write},
@@ -7,9 +10,7 @@ use std::{
 };
 
 use bytes::Bytes;
-use clap::Parser;
-use cli::{DeployerOptions, parse_private_key};
-use error::DeployerError;
+use errors::DeployerError;
 use ethrex_common::{Address, U256};
 use ethrex_l2::utils::test_data_io::read_genesis_file;
 use ethrex_l2_common::calldata::Value;
@@ -23,10 +24,15 @@ use ethrex_rpc::{
     types::block_identifier::{BlockIdentifier, BlockTag},
 };
 use keccak_hash::H256;
-use tracing::{Level, debug, error, info, trace, warn};
+use tracing::{debug, error, info, trace, warn};
 
-mod cli;
-mod error;
+use options::DeployerOptions;
+
+use crate::l2::deployer::options::parse_private_key;
+
+lazy_static::lazy_static! {
+    static ref SALT: std::sync::Mutex<H256>  = std::sync::Mutex::new(H256::zero());
+}
 
 /// Bytecode of the OnChainProposer contract.
 /// This is generated by the [build script](./build.rs).
@@ -84,45 +90,7 @@ pub struct ContractAddresses {
     pub aligned_aggregator_address: Address,
 }
 
-#[tokio::main]
-async fn main() -> Result<(), DeployerError> {
-    tracing_subscriber::fmt().with_max_level(Level::INFO).init();
-
-    trace!("Starting deployer binary");
-    let opts = DeployerOptions::parse();
-
-    let eth_client = EthClient::new_with_config(
-        vec![&opts.rpc_url],
-        opts.max_number_of_retries,
-        opts.backoff_factor,
-        opts.min_retry_delay,
-        opts.max_retry_delay,
-        Some(opts.maximum_allowed_max_fee_per_gas),
-        Some(opts.maximum_allowed_max_fee_per_blob_gas),
-    )?;
-
-    let contract_addresses = deploy_contracts(&eth_client, &opts).await?;
-
-    initialize_contracts(contract_addresses, &eth_client, &opts).await?;
-
-    if opts.deposit_rich {
-        let _ = make_deposits(contract_addresses.bridge_address, &eth_client, &opts)
-            .await
-            .inspect_err(|err| {
-                warn!("Failed to make deposits: {err}");
-            });
-    }
-
-    write_contract_addresses_to_env(contract_addresses, opts.env_file_path)?;
-    trace!("Deployer binary finished successfully");
-    Ok(())
-}
-
-lazy_static::lazy_static! {
-    static ref SALT: std::sync::Mutex<H256>  = std::sync::Mutex::new(H256::zero());
-}
-
-async fn deploy_contracts(
+pub async fn deploy_contracts(
     eth_client: &EthClient,
     opts: &DeployerOptions,
 ) -> Result<ContractAddresses, DeployerError> {
@@ -302,7 +270,7 @@ fn read_vk(path: &str) -> Bytes {
     }).into()
 }
 
-async fn initialize_contracts(
+pub async fn initialize_contracts(
     contract_addresses: ContractAddresses,
     eth_client: &EthClient,
     opts: &DeployerOptions,
@@ -495,7 +463,7 @@ async fn initialize_contracts(
     Ok(())
 }
 
-async fn make_deposits(
+pub async fn make_deposits(
     bridge: Address,
     eth_client: &EthClient,
     opts: &DeployerOptions,
@@ -573,7 +541,7 @@ async fn make_deposits(
     Ok(())
 }
 
-fn write_contract_addresses_to_env(
+pub fn write_contract_addresses_to_env(
     contract_addresses: ContractAddresses,
     env_file_path: Option<PathBuf>,
 ) -> Result<(), DeployerError> {
