+++ cmd/ethrex/l2/system_contracts_updater/errors.rs
