+++ cmd/ethrex/l2/system_contracts_updater/mod.rs
@@ -1,18 +1,16 @@
-use std::{collections::HashMap, path::Path, path::PathBuf};
+mod errors;
+pub mod options;
 
 use bytes::Bytes;
-use clap::Parser;
-use cli::SystemContractsUpdaterOptions;
-use error::SystemContractsUpdaterError;
+use errors::SystemContractsUpdaterError;
 use ethrex_common::types::{Genesis, GenesisAccount};
 use ethrex_common::{Address, H160, U256};
 use ethrex_l2::utils::test_data_io::read_genesis_file;
 use ethrex_l2_sdk::{
-    COMMON_BRIDGE_L2_ADDRESS, L2_TO_L1_MESSENGER_ADDRESS, address_to_word, get_erc1967_slot,
+    address_to_word, get_erc1967_slot, COMMON_BRIDGE_L2_ADDRESS, L2_TO_L1_MESSENGER_ADDRESS,
 };
 use genesis_tool::genesis::write_genesis_as_json;
-mod cli;
-mod error;
+use std::{collections::HashMap, path::Path, path::PathBuf};
 
 /// Bytecode of the CommonBridgeL2 contract.
 /// This is generated by the [build script](./build.rs).
@@ -48,12 +46,6 @@ pub const IMPL_MASK: Address = H160([
     0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
     0x00, 0x00, 0x10, 0x00,
 ]);
-
-fn main() -> Result<(), SystemContractsUpdaterError> {
-    let opts = SystemContractsUpdaterOptions::parse();
-    update_genesis_file(&opts.l2_genesis_path)?;
-    Ok(())
-}
 
 fn add_with_proxy(
     genesis: &mut Genesis,
@@ -93,7 +85,7 @@ fn add_with_proxy(
     Ok(())
 }
 
-fn update_genesis_file(l2_genesis_path: &PathBuf) -> Result<(), SystemContractsUpdaterError> {
+pub fn update_genesis_file(l2_genesis_path: &PathBuf) -> Result<(), SystemContractsUpdaterError> {
     let mut genesis = read_genesis_file(l2_genesis_path.to_str().ok_or(
         SystemContractsUpdaterError::InvalidPath(
             "Failed to convert l2 genesis path to string".to_string(),
