--- crates/l2/contracts/.gitignore
@@ -1,5 +1,0 @@
-# Dotenv file
-.env
-
-# Libs
-lib/
